pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "signalhub"
version = "0.1.0"
description = "SMTP to Pushover notification bridge"
authors = [ { name = "SignalHub Maintainers" } ]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "passlib[bcrypt]",
    "python-jose",
    "python-multipart",
    "pyyaml",
    "aiohttp",
    "aiosmtpd",
    "bcrypt==4.0.1",
    "cryptography",
    "email-validator"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
signalhub = "signalhub.app:main"
